pyproject.toml
[project]
name = "deploy-agent"
version = "0.2.0"
description = "AI-powered Linux server deployment agent using LangGraph"
requires-python = ">=3.11"
dependencies = [
    # AI / LangChain
    "langgraph>=0.4",
    "langchain-google-genai>=1.0",
    "langchain-ollama>=0.3",
    "langchain-openai>=0.3",
    "langchain-core>=0.3",

    # SSH
    "paramiko>=3.5",

    # Database
    "psycopg2-binary>=2.9",

    # CLI
    "typer[all]>=0.15",
    "rich>=13.0",

    # Config
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",

    # MCP Server
    "mcp[cli]>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.10",
]

[project.scripts]
deploy-ai = "src.cli.app:app"
deploy-ai-mcp = "src.mcp_server.server:mcp.run"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

# ──────────────────────────────────────────────
# Ruff — linting & formatting
# ──────────────────────────────────────────────

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort (import sorting)
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",   # line length (handled by formatter)
    "B008",   # function call in default argument (needed for Typer/FastAPI)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ──────────────────────────────────────────────
# Mypy — type checking
# ──────────────────────────────────────────────

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Gradual adoption
ignore_missing_imports = true

# ──────────────────────────────────────────────
# Pytest
# ──────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
